+++ tests/backtesting/test_portfolio.py
@@ -5,7 +5,7 @@ import pytest
 from pytest_cases import parametrize_with_cases
 
 from ptahlmud.backtesting.portfolio import Portfolio, WealthItem, WealthSeries
-from ptahlmud.backtesting.trading import Position, Trade
+from ptahlmud.backtesting.position import Position, Trade
 from ptahlmud.types.signal import Side
